+++ R/createCluster.R
@@ -1,23 +1,34 @@
-#' Create a thermal cluster
+
+#' @title Create a cluster
+#' 
+#' @description Create a new thermal or RES (renewable energy source) cluster.
 #'
 #' @param area The area where to create the cluster.
-#' @param cluster_name cluster name.
+#' @param cluster_name Name for the cluster, it will prefixed by area name, unless you set `add_prefix = FALSE`.
+#' @param group Group of the cluster, depends on cluster type:
+#'  - thermal cluster, one of: Gas, Hard coal, Lignite, Mixed fuel, Nuclear, Oil, Other, Other 2, Other 3, Other 4.
+#'  - renewable cluster, one of: Wind Onshore, Wind Offshore, Solar Thermal, Solar PV, Solar Rooftop, Other RES 1, Other RES 2, Other RES 3, Other RES 4.
 #' @param ... Parameters to write in the Ini file. Careful!
-#'  Some parameters must be set as \code{integers} to avoid warnings in Antares, for example, 
-#'  to set \code{unitcount}, you'll have to use \code{unitcount = 1L}.
+#'  Some parameters must be set as `integers` to avoid warnings in Antares, for example, 
+#'  to set \code{unitcount}, you'll have to use `unitcount = 1L`.
 #' @param time_series the "ready-made" 8760-hour time-series available for simulation purposes.
-#' @param prepro_data Pre-process data, a \code{data.frame} or \code{matrix}, 
+#' @param prepro_data Pre-process data, a `data.frame` or `matrix`, 
 #'  default is a matrix with 365 rows and 6 columns.
-#' @param prepro_modulation Pre-process modulation, a \code{data.frame} or \code{matrix}, 
+#' @param prepro_modulation Pre-process modulation, a `data.frame` or `matrix`, 
 #'  if specified, must have 8760 rows and 1 or 4 columns.
-#' @param add_prefix If \code{TRUE}, cluster_name will be prefixed by area name.
+#' @param add_prefix If `TRUE` (the default), `cluster_name` will be prefixed by area name.
 #' @param overwrite Logical, overwrite the cluster or not.
 #' @param opts
 #'   List of simulation parameters returned by the function
-#'   \code{antaresRead::setSimulationPath}
+#'  [antaresRead::setSimulationPath()]
 #'
 #' @return An updated list containing various information about the simulation.
+#' 
+#' @seealso [editCluster()] or [editClusterRES()] to edit existing clusters, [removeCluster()] or [removeClusterRES()] to remove clusters.
+#' 
 #' @export
+#' 
+#' @name create-cluster
 #' 
 #' @importFrom antaresRead simOptions
 #' @importFrom stats setNames
@@ -36,7 +47,7 @@
 #'   cluster_name = "my_cluster",
 #'   group = "other", 
 #'   unitcount = 1L, # or as.integer(1)
-#'   `marginal-cost` = 50
+#'   marginal_cost = 50
 #' )
 #' # by default, cluster name is prefixed 
 #' # by the area name
@@ -49,7 +60,7 @@
 #'   cluster_name = "my_cluster",
 #'   add_prefix = FALSE,
 #'   group = "other", 
-#'   `marginal-cost` = 50
+#'   marginal_cost = 50
 #' )
 #' levels(readClusterDesc()$cluster)
 #' # > "my_cluster"
@@ -108,31 +119,121 @@
 #' )
 #' 
 #' }
-createCluster <- function(area, cluster_name, ..., time_series = NULL,
-                          prepro_data = NULL, prepro_modulation = NULL,
-                          add_prefix = TRUE, overwrite = FALSE,
+createCluster <- function(area, 
+                          cluster_name, 
+                          group = "Other",
+                          ...,
+                          time_series = NULL,
+                          prepro_data = NULL,
+                          prepro_modulation = NULL,
+                          add_prefix = TRUE, 
+                          overwrite = FALSE,
                           opts = antaresRead::simOptions()) {
+  thermal_group <- c("Gas",
+                     "Hard coal",
+                     "Lignite",
+                     "Mixed fuel",
+                     "Nuclear",
+                     "Oil",
+                     "Other",
+                     "Other 2",
+                     "Other 3",
+                     "Other 4")
+  if (!is.null(group) && !group %in% thermal_group) 
+    warning(
+      "Group: '", group, "' is not a valid name recognized by Antares,",
+      " you should be using one of: ", paste(thermal_group, collapse = ", ")
+    )
+  .createCluster(
+    area = area, 
+    cluster_name = cluster_name, 
+    group = group,
+    ...,
+    time_series = time_series,
+    prepro_data = prepro_data,
+    prepro_modulation = prepro_modulation,
+    add_prefix = add_prefix, 
+    overwrite = overwrite,
+    cluster_type = "thermal",
+    opts = opts
+  )
+}
+
+#' @export
+#' 
+#' @rdname create-cluster
+createClusterRES <- function(area, 
+                             cluster_name, 
+                             group = "Other RES 1",
+                             ...,
+                             time_series = NULL,
+                             add_prefix = TRUE, 
+                             overwrite = FALSE,
+                             opts = antaresRead::simOptions()) {
+  assertthat::assert_that(class(opts) == "simOptions")
+  check_active_RES(opts)
+  renewables_group <- c("Wind Onshore",
+                        "Wind Offshore",
+                        "Solar Thermal",
+                        "Solar PV",
+                        "Solar Rooftop",
+                        "Other RES 1",
+                        "Other RES 2",
+                        "Other RES 3",
+                        "Other RES 4")
+  if (!is.null(group) && !group %in% renewables_group) 
+    warning(
+      "Group: '", group, "' is not a valid name recognized by Antares,",
+      " you should be using one of: ", paste(renewables_group, collapse = ", ")
+    )
+  initialize_RES(opts)
+  .createCluster(
+    area = area, 
+    cluster_name = cluster_name,
+    group = group,
+    ...,
+    time_series = time_series,
+    prepro_data = NULL,
+    prepro_modulation = NULL,
+    add_prefix = add_prefix, 
+    overwrite = overwrite,
+    cluster_type = "renewables",
+    opts = opts
+  )
+}
+
+
+.createCluster <- function(area, 
+                           cluster_name, 
+                           ...,
+                           time_series = NULL,
+                           prepro_data = NULL,
+                           prepro_modulation = NULL,
+                           add_prefix = TRUE, 
+                           overwrite = FALSE,
+                           cluster_type = c("thermal", "renewables"),
+                           opts = antaresRead::simOptions()) {
 
   # Input path
   inputPath <- opts$inputPath
   assertthat::assert_that(!is.null(inputPath) && file.exists(inputPath))
+  cluster_type <- match.arg(cluster_type)
   
-  if (!tolower(area) %in% opts$areaList)
-    stop(paste(area, "is not a valid area"))
+  check_area_name(area, opts)
   
-  if (! NROW(time_series) %in% c(0, 8736, 8760)) {
+  if (!NROW(time_series) %in% c(0, 8736, 8760)) {
     stop("Number of rows for time series must be 0 or 8760")
   }
   
-  if (! NROW(prepro_modulation) %in% c(0, 8736, 8760)) {
+  if (!NROW(prepro_modulation) %in% c(0, 8736, 8760)) {
     stop("Number of rows for modulation data must be 0 or 8760")
   }
-  if (! NCOL(prepro_modulation) %in% c(1, 4)) {
+  if (!NCOL(prepro_modulation) %in% c(1, 4)) {
     stop("Number of cols for modulation data must be 0 or 4")
   }
 
   # Cluster's parameters
-  params_cluster <- list(...)
+  params_cluster <- hyphenize_names(list(...))
   if (add_prefix)
     cluster_name <- paste(area, cluster_name, sep = "_")
   params_cluster$name <- cluster_name
@@ -140,8 +241,8 @@ createCluster <- function(area, cluster_name, ..., time_series = NULL,
   # named list for writing ini file
   # params_cluster <- stats::setNames(object = list(params_cluster), nm = cluster_name)
 
-  # path to ini file
-  path_clusters_ini <- file.path(inputPath, "thermal", "clusters", tolower(area), "list.ini")
+  # path to ini file containing clusters' name and parameters
+  path_clusters_ini <- file.path(inputPath, cluster_type, "clusters", tolower(area), "list.ini")
 
   # read previous content of ini
   previous_params <- readIniFile(file = path_clusters_ini)
@@ -167,7 +268,7 @@ createCluster <- function(area, cluster_name, ..., time_series = NULL,
 
   # initialize series
   dir.create(path = file.path(
-    inputPath, "thermal", "series", tolower(area), tolower(cluster_name)),
+    inputPath, cluster_type, "series", tolower(area), tolower(cluster_name)),
     recursive = TRUE, showWarnings = FALSE
   )
   
@@ -185,40 +286,43 @@ createCluster <- function(area, cluster_name, ..., time_series = NULL,
   
   utils::write.table(
     x = time_series, row.names = FALSE, col.names = FALSE, sep = "\t",
-    file = file.path(inputPath, "thermal", "series", tolower(area), tolower(cluster_name), "series.txt")
+    file = file.path(inputPath, cluster_type, "series", tolower(area), tolower(cluster_name), "series.txt")
   )
 
 
   # prepro
-  dir.create(
-    path = file.path(inputPath, "thermal", "prepro", tolower(area), tolower(cluster_name)),
-    recursive = TRUE, showWarnings = FALSE
-  )
-  
-  if (is.null(prepro_data))
-    prepro_data <- matrix(data = c(rep(1, times = 365 * 2), rep(0, times = 365 * 4)), ncol = 6)
-  utils::write.table(
-    x = prepro_data, row.names = FALSE, col.names = FALSE, sep = "\t",
-    file = file.path(inputPath, "thermal", "prepro", tolower(area), tolower(cluster_name), "data.txt")
-  )
-  
-  
-  if (is.null(prepro_modulation))
-    prepro_modulation <- matrix(data = c(rep(1, times = 365 * 24 * 3), rep(0, times = 365 * 24 * 1)), ncol = 4)
-  
-  utils::write.table(
-    x = prepro_modulation, row.names = FALSE, col.names = FALSE, sep = "\t",
-    file = file.path(inputPath, "thermal", "prepro", tolower(area), tolower(cluster_name), "modulation.txt")
-  )
+  if (identical(cluster_type, "thermal")) {
+    dir.create(
+      path = file.path(inputPath, cluster_type, "prepro", tolower(area), tolower(cluster_name)),
+      recursive = TRUE, showWarnings = FALSE
+    )
+    
+    if (is.null(prepro_data))
+      prepro_data <- matrix(data = c(rep(1, times = 365 * 2), rep(0, times = 365 * 4)), ncol = 6)
+    utils::write.table(
+      x = prepro_data, row.names = FALSE, col.names = FALSE, sep = "\t",
+      file = file.path(inputPath, cluster_type, "prepro", tolower(area), tolower(cluster_name), "data.txt")
+    )
+    
+    
+    if (is.null(prepro_modulation))
+      prepro_modulation <- matrix(data = c(rep(1, times = 365 * 24 * 3), rep(0, times = 365 * 24 * 1)), ncol = 4)
+    
+    utils::write.table(
+      x = prepro_modulation, row.names = FALSE, col.names = FALSE, sep = "\t",
+      file = file.path(inputPath, cluster_type, "prepro", tolower(area), tolower(cluster_name), "modulation.txt")
+    )
+  }
 
 
-  # Maj simulation
+  # Update simulation options object
   suppressWarnings({
     res <- antaresRead::setSimulationPath(path = opts$studyPath, simulation = "input")
   })
   
   invisible(res)
 }
+